backend/packages/pyproject.toml
[tool.poetry]
name = "websearch"
version = "0.1.0"
description = "This package for support agents to find KB and process faster."
authors = ["Paphaphin Kaewchankha <p.a.kaewchankha@accenture.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
anthropic = "0.34.2"
fastapi = "0.115.0"
fastapi-cli = "0.0.5"
httpx = "0.27.2"
langchain = "0.3.1"
langchain-anthropic = "0.2.1"
langchain-community = "0.3.0"
langchain-core = "0.3.6"
langchain-experimental = "0.3.0"
langchain-google-vertexai = "2.0.3"
langdetect = "1.0.9"
langgraph = "0.2.26"
langgraph-checkpoint = "1.0.11"
langserve = "0.3.0"
langsmith = "0.1.126"
python-dotenv = "1.0.1"
passlib = "*"
uvicorn = "0.23.2"
